requirements.txt
tabulate==0.8.7
requests==2.25.0
pandas==1.1.4
xlrd==1.2.0
more-itertools==8.6.0
PyYAML==5.3.1
click==7.1.2
